requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.60.1
fsspec==2026.1.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
opt_einsum==3.4.0
packaging==25.0
patsy==1.0.2
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==80.10.2
sgmllib3k==1.0.0
six==1.17.0
starlette==0.50.0
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0

numpy
torchvision
transformers
huggingface-hub
safetensors

